Struttura riorganizzata.
Ho riorganizzato (ancora) la struttura del progetto e pulito un po' il
codice, smistando alcuni metodi tra delle nuove classi, soprattutto per
la view. Poi ho sistemato la struttura MVC perché alcuni metodi del
controller svolgevano troppe elaborazioni.
