+++ src/main/gui/GuiManager.java
@@ -1,0 +1,11 @@
+package main.gui;
+
+/**
+ * Main GUI class which sets up the graphical appearance and manages any kind of interaction with the user.
+ * It is also the interface for the communications between the view module and the controller.
+ *
+ * @author Manuel Gallina
+ */
+public interface GuiManager {
+    void start();
+}
