+++ src/main/gui/screens/OperatorScreen.java
@@ -1,0 +1,38 @@
+package main.gui.screens;
+
+import main.SystemController;
+import main.utility.InputParserUtility;
+
+import java.util.logging.Logger;
+
+import static main.utility.Notifications.PROMPT_OPERATOR_CHOICES;
+
+/**
+ * The operator menu screen.
+ *
+ * @author Manuel Gallina
+ */
+class OperatorScreen extends Screen {
+    private Logger logger = Logger.getLogger(this.getClass().getName());
+
+    public OperatorScreen(SystemController controller) {
+        super(controller);
+
+        System.out.printf("%s\n\n", PROMPT_OPERATOR_CHOICES);
+
+        String command;
+        do {
+            command = getScanner().nextLine();
+        } while(!InputParserUtility.isValidInteger(command, 1, 3));
+
+        switch(Integer.parseInt(command)) {
+            case 1:
+                System.out.println(controller.allUsersToString());
+                break;
+            case 2://logout
+                break;
+            default:
+                break;
+        }
+    }
+}
