+++ src/main/model/user/UserStatus.java
@@ -1,0 +1,11 @@
+package main.model.user;
+
+/**
+ * The user status.
+ *
+ * @author Manuel Gallina
+ */
+public enum UserStatus {
+    CUSTOMER,
+    OPERATOR;
+}
